go: Use JWKS for key rotation
Instead of trying all possible pub keys one-by-one, build a map of kid
(fingerprint) to pub key and use the kid header in jwt to select the key
to use. If either the kid header is absent or we don't have any key with
matching fingerprint, fallback to the first (usually current) key from
the versioned secret.

This also allows us to switch from reddit's forked version of jwt back
to the upstream version.