allow edgecontext to wrap other implementations (#28)
This PR extracts the actual implementation of the EdgeRequestContext into an interface which allows us to plug in our internal, v2 implementation. This means that services that have code that use both this library and the v2 implementation will share the same underlying objects, avoiding multiple deserialization and verification calls.